% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_controls_dw.r
\docType{data}
\name{cps_controls_dw}
\alias{cps_controls_dw}
\title{CPS control data used in the Dehejia and Wahba (1999) benchmarking
of Lalonde's (1986) original study.}
\format{
A data frame with 18361 rows and 12 variables:
\describe{
  \item{re75}{real earnings in 1975 (using 1975 dollars)}
  \item{re74}{real earnings in 1975 (using 1975 dollars)}
  \item{treat}{binary indicator of enrollment in NSW program}
  ...
}
}
\source{
\url{https://users.nber.org/~rdehejia/data/.nswdata2.html}
}
\usage{
cps_controls_dw
}
\description{
#'
}
\keyword{datasets}
