% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lalonde.r
\docType{data}
\name{lalonde}
\alias{lalonde}
\title{Original data from the Lalonde (1986) study.}
\format{
A data frame with 2675 rows and 12 variables:
\describe{
  \item{re75}{real earnings in 1975 (using 1975 dollars)}
  \item{re74}{real earnings in 1975 (using 1975 dollars)}
  \item{nsw}{binary indicator of enrollment in NSW program}
  ...
}
}
\source{
\url{https://users.nber.org/~rdehejia/data/.nswdata2.html}
}
\usage{
lalonde
}
\description{
A dataset containing the data from Lalonde's (1986) original NSW study.
}
\keyword{datasets}
