% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsw_dehejia_wahba.r
\docType{data}
\name{nsw_dehejia_wahba}
\alias{nsw_dehejia_wahba}
\title{Data from the Dehejia and Wahba (1999) observational benchmarking
 study of Lalonde's (1986) experimental NSW study.}
\format{
A data frame with 445 rows and 11 variables:
\describe{
  \item{re75}{real earnings in 1975 (using 1975 dollars)}
  \item{re74}{real earnings in 1975 (using 1975 dollars)}
  \item{treat}{binary indicator of enrollment in NSW program}
  ...
}
}
\source{
\url{https://users.nber.org/~rdehejia/data/.nswdata2.html}
}
\usage{
nsw_dehejia_wahba
}
\description{
Data from the Dehejia and Wahba (1999) observational benchmarking
 study of Lalonde's (1986) experimental NSW study.
}
\keyword{datasets}
